* Add options to avoid actual bucket deletion (#241)

* Rds collector (#242)

* Initial Draft of RDS Collector

* Adding missing rds_role ConfigOption

* Adjusting based on feedback

* Correcting typo in dbconfig (#244)

* Fixing account_id type and default S3 delete setting (#245)

* Fixing account_id type and default S3 delete setting

* including type change

* Fixing to be int not octal literal

* Modifying RDS based on testing

* Test fixes (#247)

* Fixing account_id type and default S3 delete setting

* including type change

* Fixing to be int not octal literal

* Modifying RDS based on testing

* Fixing typo

* Prepare for RDS auditor (#249)

* Bugfixes (#251)

* Improvements (#252)

* 1. Tests for Lambda based RDS auditor
2. Bugfixes

* 1. Tests for Lambda based RDS auditor
2. Bugfixes

* 1. Tests for Lambda based RDS auditor
2. Bugfixes

* Minor fixes (#253)

* Minor fixes

* RDS Fixes

* Missing a save

* Logging adjustments

* fixes

* Optimizations (#254)

* 1. Tests for Lambda based RDS auditor
2. Bugfixes

* 1. Tests for Lambda based RDS auditor
2. Bugfixes

* 1. Tests for Lambda based RDS auditor
2. Bugfixes

* Optimizations

* Add 'engine' param in test

* Many fixes per testing (#255)

* fix case (#256)

* Rds collector fix2 (#257)

* fix lambda payload decoding (#258)

* Made max-age an optional tag and moved GDPR constants to config (#259)

* Bugfix for auditor (#260)

* bugfixes (#261)

* Update providers.py (#262)

* bugfix (#263)

* bugfix (#264)

* bugfix (#265)

* bugfixes and improvements (#266)

* Fix duplicate stop notices (#268)

* Fixing indent causing removal of removed RDS databases to fail (#269)

* Log to Enforcement iff action succeed (#271)

+++ plugins/public/cinq-auditor-required-tags/cinq_auditor_required_tags/providers.py
@@ -3,10 +3,9 @@ import logging
 from datetime import datetime, timedelta
 
 from cinq_auditor_required_tags.utils import s3_removal_policy_exists, s3_removal_lifecycle_policy_exists
-from cloud_inquisitor.constants import ActionStatus
-
 from cloud_inquisitor import get_aws_session
 from cloud_inquisitor.config import dbconfig
+from cloud_inquisitor.constants import ActionStatus
 from cloud_inquisitor.constants import AuditActions, NS_AUDITOR_REQUIRED_TAGS
 from cloud_inquisitor.log import auditlog
 from cloud_inquisitor.plugins.types.accounts import AWSAccount
@@ -14,6 +13,11 @@ from cloud_inquisitor.plugins.types.enforcements import Enforcement
 from cloud_inquisitor.plugins.types.resources import EC2Instance
 
 logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
+
+
+def noop(client, resource):
+    return ActionStatus.SUCCEED, resource.metrics()
 
 
 def process_action(resource, action, action_issuer='unknown'):
@@ -26,18 +30,33 @@ def process_action(resource, action, action_issuer='unknown'):
     Returns:
         `ActionStatus`
     """
+    from cinq_collector_aws import AWSRegionCollector
+
     func_action = action_mapper[resource.resource_type][action]
+    extra_info = {}
+    action_status = ActionStatus.UNKNOWN
+
     if func_action:
-        client = get_aws_session(AWSAccount(resource.account)).client(
-            action_mapper[resource.resource_type]['service_name'],
-            region_name=resource.location
-        )
+        if action_mapper[resource.resource_type]['service_name'] == 'lambda':
+            client = get_aws_session(
+                AWSAccount.get(dbconfig.get('rds_collector_account', AWSRegionCollector.ns, ''))
+            ).client(
+                'lambda',
+                dbconfig.get('rds_collector_region', AWSRegionCollector.ns, '')
+            )
+        else:
+            client = get_aws_session(AWSAccount(resource.account)).client(
+                action_mapper[resource.resource_type]['service_name'],
+                region_name=resource.location
+            )
         try:
-            action_status, metrics = func_action(client, resource)
-            Enforcement.create(resource.account.account_name, resource.id, action, datetime.now(), metrics)
+            logger.info(f'Trying to {action} resource {resource.id} for account {resource.account.account_name} / region {resource.location}')
+            action_status, extra_info = func_action(client, resource)
+            if action_status == ActionStatus.SUCCEED:
+                Enforcement.create(resource.account.account_id, resource.id, action, datetime.now(), extra_info)
         except Exception as ex:
             action_status = ActionStatus.FAILED
-            logger.error('Failed to apply action {} to {}: {}'.format(action, resource.id, ex))
+            logger.exception('Failed to apply action {} to {}: {}'.format(action, resource.id, ex))
         finally:
             auditlog(
                 event='{}.{}.{}.{}'.format(action_issuer, resource.resource_type, action, action_status),
@@ -45,7 +64,8 @@ def process_action(resource, action, action_issuer='unknown'):
                 data={
                     'resource_id': resource.id,
                     'account_name': resource.account.account_name,
-                    'location': resource.location
+                    'location': resource.location,
+                    'info': extra_info
                 }
             )
             return action_status
@@ -172,8 +192,45 @@ def delete_s3_bucket(client, resource):
         `ActionStatus`
     """
 
-    client.delete_bucket(Bucket=resource.id)
+    if dbconfig.get('enable_delete_s3_buckets', NS_AUDITOR_REQUIRED_TAGS, False):
+        client.delete_bucket(Bucket=resource.id)
     return ActionStatus.SUCCEED, resource.metrics()
+
+
+def stop_rds_instance(client, resource):
+    return operate_rds_instance(client, resource, 'stop')
+
+
+def terminate_rds_instance(client, resource):
+    return operate_rds_instance(client, resource, 'terminate')
+
+
+def operate_rds_instance(client, resource, action):
+    resource_info = {
+        'platform': 'AWS',
+        'accountId': AWSAccount(resource.account).account_number,
+        'accountName': resource.account.account_name,
+        'action': action,
+        'region': resource.location,
+        'resourceId': resource.id,
+        'resourceType': 'rds',
+        'resourceSubType': resource.engine
+    }
+
+    logger.info('Auditor Request Payload: {}'.format(resource_info))
+
+    response = client.invoke(
+        FunctionName=dbconfig.get('action_taker_arn', NS_AUDITOR_REQUIRED_TAGS, ''),
+        Payload=json.dumps(resource_info).encode('utf-8')
+    )
+
+    response_payload = json.loads(response['Payload'].read().decode('utf-8'))
+    logger.info('Server Response Payload: {}'.format(response_payload))
+
+    if response_payload.get('success'):
+        return ActionStatus.SUCCEED, response_payload.get('data', {})
+    else:
+        return ActionStatus.FAILED, {'message': response_payload.get('message')}
 
 
 action_mapper = {
@@ -186,5 +243,10 @@ action_mapper = {
         'service_name': 's3',
         AuditActions.STOP: stop_s3_bucket,
         AuditActions.REMOVE: delete_s3_bucket
+    },
+    'aws_rds_instance': {
+        'service_name': 'lambda',
+        AuditActions.STOP: stop_rds_instance,
+        AuditActions.REMOVE: terminate_rds_instance
     }
 }
